+++ test/test_add_points_in_pointcloud.py
@@ -2,151 +2,96 @@ import inspect
 import os
 from pathlib import Path
 
-import pdal
+import geopandas as gpd
 import pytest
 
 from pdaltools import add_points_in_pointcloud
+from pdaltools.count_occurences.count_occurences_for_attribute import (
+    compute_count_one_file,
+)
 
 TEST_PATH = os.path.dirname(os.path.abspath(__file__))
-TMP_PATH = os.path.join(TEST_PATH, "data/output")
+TMP_PATH = os.path.join(TEST_PATH, "tmp/add_points_in_pointcloud")
 DATA_LIDAR_PATH = os.path.join(TEST_PATH, "data/decimated_laz")
 DATA_POINTS_PATH = os.path.join(TEST_PATH, "data/points_3d")
 
-INPUT_FILE = os.path.join(DATA_LIDAR_PATH, "test_semis_2023_0292_6833_LA93_IGN69.laz")
+INPUT_PCD = os.path.join(DATA_LIDAR_PATH, "test_semis_2023_0292_6833_LA93_IGN69.laz")
 INPUT_POINTS = os.path.join(DATA_POINTS_PATH, "Points_virtuels_0292_6833.geojson")
 OUTPUT_FILE = os.path.join(TMP_PATH, "test_semis_2023_0292_6833_LA93_IGN69.laz")
 
-INPUT_FILE_SMALL = os.path.join(DATA_LIDAR_PATH, "test_semis_2021_0382_6565_LA93_IGN69.laz")
-INPUT_POINTS_SMALL = os.path.join(DATA_POINTS_PATH, "Points_virtuels_0382_6565.geojson")
-OUTPUT_FILE_SMALL = os.path.join(TMP_PATH, "test_semis_2021_0382_6565_LA93_IGN69.laz")
+# Cropped las tile used to test adding points that belong to the theorical tile but not to the
+# effective las file extent
+INPUT_PCD_CROPPED = os.path.join(DATA_LIDAR_PATH, "test_semis_2021_0382_6565_LA93_IGN69_cropped.laz")
+INPUT_POINTS_FOR_CROPPED_PCD = os.path.join(DATA_POINTS_PATH, "Points_virtuels_0382_6565.geojson")
+OUTPUT_FILE_CROPPED_PCD = os.path.join(TMP_PATH, "test_semis_2021_0382_6565_LA93_IGN69.laz")
 
 
 def setup_module(module):
-    os.makedirs("test/data/output", exist_ok=True)
+    os.makedirs(TMP_PATH, exist_ok=True)
 
 
-def test_get_tile_bbox():
-    bbox = add_points_in_pointcloud.get_tile_bbox(INPUT_FILE, 1000)
-    assert bbox == (292000.0, 6832000.0, 293000.0, 6833000.0)  # check the bbox from LIDAR tile
-
-
-def test_clip_3d_points_to_tile():
+@pytest.mark.parametrize(
+    "epsg",
+    [
+        "EPSG:2154",  # should work when providing an epsg value
+        None,  # Should also work with no epsg value (get from las file)
+    ],
+)
+def test_clip_3d_points_to_tile(epsg):
     # With EPSG
-    points_clipped = add_points_in_pointcloud.clip_3d_points_to_tile(INPUT_POINTS, INPUT_FILE, "EPSG:2154", 1000)
-    assert len(points_clipped) == 678  # chech the entity's number of points
+    points_clipped = add_points_in_pointcloud.clip_3d_points_to_tile(INPUT_POINTS, INPUT_PCD, epsg, 1000)
+    assert len(points_clipped) == 678  # check the entity's number of points
 
 
-def test_clip_3d_points_to_tile_from_epsg_none():
-    points_clipped = add_points_in_pointcloud.clip_3d_points_to_tile(INPUT_POINTS, INPUT_FILE, None, 1000)
-    assert len(points_clipped) == 678  # chech the entity's number of points
-
-
-def test_add_line_to_lidar():
+@pytest.mark.parametrize(
+    "input_file, epsg, expected_nb_points",
+    [
+        (INPUT_PCD, "EPSG:2154", 2423),  # should work when providing an epsg value
+        (INPUT_PCD, None, 2423),  # Should also work with no epsg value (get from las file)
+        (INPUT_PCD_CROPPED, None, 2423),
+    ],
+)
+def test_add_points_to_las(input_file, epsg, expected_nb_points):
     # Ensure the output file doesn't exist before the test
     if Path(OUTPUT_FILE).exists():
         os.remove(OUTPUT_FILE)
 
-    points_clipped = add_points_in_pointcloud.clip_3d_points_to_tile(INPUT_POINTS, INPUT_FILE, "EPSG:2154", 1000)
-
-    add_points_in_pointcloud.add_points_to_las(points_clipped, INPUT_FILE, OUTPUT_FILE, "EPSG:2154", 68)
+    points = gpd.read_file(INPUT_POINTS)
+    add_points_in_pointcloud.add_points_to_las(points, input_file, OUTPUT_FILE, epsg, 68)
     assert Path(OUTPUT_FILE).exists()  # check output exists
 
-    # Filter pointcloud by classes
-    pipeline = (
-        pdal.Reader.las(filename=OUTPUT_FILE, nosrs=True)
-        | pdal.Filter.range(
-            limits="Classification[68:68]",
-        )
-        | pdal.Filter.stats()
-    )
-    pipeline.execute()
-    metadata = pipeline.metadata
-    # Count the pointcloud's number from classe "68"
-    point_count = metadata["metadata"]["filters.stats"]["statistic"][0]["count"]
-    assert point_count == 678
+    point_count = compute_count_one_file(OUTPUT_FILE)["68"]
+    assert point_count == expected_nb_points  # Add all points from geojson
 
 
-def test_add_line_to_lidar_from_epsg_none():
+@pytest.mark.parametrize(
+    "input_file, input_points, epsg, expected_nb_points",
+    [
+        (INPUT_PCD, INPUT_POINTS, None, 678),  # should add only points within tile extent
+        (INPUT_PCD_CROPPED, INPUT_POINTS_FOR_CROPPED_PCD, None, 186),
+        (
+            INPUT_PCD_CROPPED,
+            INPUT_POINTS,
+            None,
+            0,
+        ),  # Should add no points when there is only points outside the tile extent
+        (
+            INPUT_PCD_CROPPED,
+            INPUT_POINTS_FOR_CROPPED_PCD,
+            "EPSG:2154",
+            186,
+        ),  # Should work with or without an input epsg
+    ],
+)
+def test_add_points_from_geojson_to_las(input_file, input_points, epsg, expected_nb_points):
     # Ensure the output file doesn't exist before the test
     if Path(OUTPUT_FILE).exists():
         os.remove(OUTPUT_FILE)
 
-    points_clipped = add_points_in_pointcloud.clip_3d_points_to_tile(INPUT_POINTS, INPUT_FILE, None, 1000)
-
-    add_points_in_pointcloud.add_points_to_las(points_clipped, INPUT_FILE, OUTPUT_FILE, None, 68)
+    add_points_in_pointcloud.add_points_from_geojson_to_las(input_points, input_file, OUTPUT_FILE, 68, epsg, 1000)
     assert Path(OUTPUT_FILE).exists()  # check output exists
-
-    # Filter pointcloud by classes
-    pipeline = (
-        pdal.Reader.las(filename=OUTPUT_FILE, nosrs=True)
-        | pdal.Filter.range(
-            limits="Classification[68:68]",
-        )
-        | pdal.Filter.stats()
-    )
-    pipeline.execute()
-    metadata = pipeline.metadata
-    # Count the pointcloud's number from classe "68"
-    point_count = metadata["metadata"]["filters.stats"]["statistic"][0]["count"]
-    assert point_count == 678
-
-
-def test_get_tile_bbox_small():
-    # Tile is not complete (NOT 1km * 1km)
-    bbox = add_points_in_pointcloud.get_tile_bbox(INPUT_FILE_SMALL, 1000)
-    assert bbox == (382000.0, 6564000.0, 383000.0, 6565000.0)  # return BBOX 1km * 1km
-
-
-def test_add_line_to_lidar_small():
-    # Ensure the output file doesn't exist before the test
-    if Path(OUTPUT_FILE_SMALL).exists():
-        os.remove(OUTPUT_FILE_SMALL)
-
-    # Tile is not complete (NOT 1km * 1km)
-    points_clipped = add_points_in_pointcloud.clip_3d_points_to_tile(
-        INPUT_POINTS_SMALL, INPUT_FILE_SMALL, "EPSG:2154", 1000
-    )
-
-    add_points_in_pointcloud.add_points_to_las(points_clipped, INPUT_FILE_SMALL, OUTPUT_FILE_SMALL, "EPSG:2154", 68)
-    assert Path(OUTPUT_FILE).exists()  # check output exists
-
-    # Filter pointcloud by classes
-    pipeline = (
-        pdal.Reader.las(filename=OUTPUT_FILE_SMALL, nosrs=True)
-        | pdal.Filter.range(
-            limits="Classification[68:68]",
-        )
-        | pdal.Filter.stats()
-    )
-    pipeline.execute()
-    metadata = pipeline.metadata
-    # Count the pointcloud's number from classe "68"
-    point_count = metadata["metadata"]["filters.stats"]["statistic"][0]["count"]
-    assert point_count == 186
-
-
-def test_add_points_from_geojson_to_las():
-    # Ensure the output file doesn't exist before the test
-    if Path(OUTPUT_FILE).exists():
-        os.remove(OUTPUT_FILE)
-
-    add_points_in_pointcloud.add_points_from_geojson_to_las(
-        INPUT_POINTS, INPUT_FILE, OUTPUT_FILE, 68, "EPSG:2154", 1000
-    )
-    assert Path(OUTPUT_FILE).exists()  # check output exists
-
-
-def test_add_points_from_geojson_to_las_no_epsg():
-    # Ensure the output file doesn't exist before the test
-    if Path(OUTPUT_FILE).exists():
-        os.remove(OUTPUT_FILE)
-
-    INPUT_FILE_WITHOUT_EPSG = os.path.join(TEST_PATH, "data/test_noepsg_043500_629205_IGN69.laz")
-
-    with pytest.raises(RuntimeError, match="does not have a valid EPSG code"):
-        add_points_in_pointcloud.add_points_from_geojson_to_las(
-            INPUT_POINTS, INPUT_FILE_WITHOUT_EPSG, OUTPUT_FILE, 68, None, 1000
-        )
+    point_count = compute_count_one_file(OUTPUT_FILE)["68"]
+    assert point_count == expected_nb_points  # Add all points from geojson
 
 
 def test_parse_args():